Revert "feat: move monthlychart components to src/pages/listPage/monthlyChart"
This reverts commit e9cafd79c0c7acaaac1ad150655a1978e4c76a15.
